Completed Engine class rewrite 	+ Program can be compiled and executed without crashing 	+ Ant behaviour is consistent but final results are not 	+ Things need fixing: 		* Grid size is bigger than the display 		* Ant coordinates are placed in wrong locations 		* Possible border wrapping miscalculation

+++ C++/Ant.cpp
@@ -19,7 +19,7 @@ Ant::Ant(int x, int y, int orientation, std::string rules, Tile* temp){
 	*/
 	tile = temp;
 	dir = orientation;
-	pos = glm::vec2(x, y);
+	pos = glm::ivec2(x, y);
 	ruleset = rules;
 	}
